requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
numpy>=1.22.0
scikit-learn>=1.2.0
plotly>=5.10.0
tensorflow>=2.9.0
keras-tcn>=3.5.0
